pyproject.toml
[tool.poetry]
name = "transcribe"
version = "0.1.0"
description = ""
authors = ["Matthias Osswald <info@busykoala.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pydub = "^0.25.1"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
python-multipart = "^0.0.20"
python-dotenv = "^1.0.1"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.3"
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = ["E", "F", "W", "Q", "I"]
ignore = ["E501"]

[tool.ruff.lint.isort]
force-single-line = true
